Add files via upload
The need for deep learning in the astronomical community has become increasingly apparent, as advances in data acquisition now require complex algorithms and enormous computing power. Data sets such as the Sloan Digital Sky Survey (SDSS), which alone has catalogued over 1⁄3 of the sky, are growing in size so quickly that automated tools are now necessary for mining through the overwhelming abundance of information (Sánchez et al. 2018). As classification is a shared goal of Machine Learning and Astrophysics, we need to look at the most recent advancements in the world of deep learning of the past two years.

Our primary goal is to construct a novel algorithm that can aptly classify galaxies into three known types (spiral, elliptical, and irregular) based on their observed and derived photometric properties. Our self-training classifier utilizes a semi-supervised learning technique and is based upon the DeepForest model, proposed and built by Zhou and Feng (2019). and our data source is the ALFALFA-SDSS Galaxy Catalog, which contains basic optical and derived properties of ~ 30,000 galaxies. We found that our classifier performed with 83% accuracy and an F1-score of 78%. The model’s performance supports the notion that it is possible to a) classify galaxies without having to rely on the standard methods of backpropagation and differentiation and b) using only observed/derived properties and no images. DeepForest has opened the doors to allowing new frameworks for classification in the astronomical community and beyond.